User management: complete crud Role and Permission

+++ app/Http/Controllers/Admin/UserManagements/RoleController.php
@@ -17,26 +17,33 @@ class RoleController extends Controller
 {
     public function index(Request $request)
     {
-//        abort_if(Gate::denies('role_access'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(! Gate::allows('role_access'),403);
 
         if($request->ajax())
         {
             $data=Role::query();
 
-            if (request('trash') == 1){
+            if (request('trash') == 1 && Gate::allows('role_delete')){
                 $data=$data->onlyTrashed()->get();
             }
 
             return DataTables::of($data)
                 ->addColumn('action', function ($data) {
-                    $button='<a href="' . route('admin.user_managements.roles.show', $data->id) . '" class="btn btn-sm btn-info mr-1 mb-1" data-toggle="tooltip" title="Show data"><i class="fa fa-eye"></i></a>';
-                    $button .=' <a href="' . route('admin.user_managements.roles.edit', $data->id) . '" class="btn btn-sm btn-success mr-1 mb-1" data-toggle="tooltip" title="Edit data"><i class="fa fa-edit"></i></a>';
-                    $button .=' <button type="button" name="delete" id="'.$data->id.'" class="btn btn-sm btn-danger mr-1 mb-1 delete" data-toggle="tooltip" title="Delete data"><i class="fa fa-trash-alt"></i></button>';
+                    $button='';
+                    if (Gate::allows('role_show')){
+                        $button .='<a href="' . route('admin.user_managements.roles.show', $data->id) . '" class="btn btn-sm btn-info mr-1 mb-1" data-toggle="tooltip" title="Show data"><i class="fa fa-eye"></i></a>';
+                    }
+                    if (Gate::allows('role_update')){
+                        $button .=' <a href="' . route('admin.user_managements.roles.edit', $data->id) . '" class="btn btn-sm btn-success mr-1 mb-1" data-toggle="tooltip" title="Edit data"><i class="fa fa-edit"></i></a>';
+                    }
+                    if (Gate::allows('role_delete')){
+                        $button .=' <button type="button" name="delete" id="'.$data->id.'" class="btn btn-sm btn-danger mr-1 mb-1 delete" data-toggle="tooltip" title="Delete data"><i class="fa fa-trash-alt"></i></button>';
+                    }
 
                     $trash =' <button type="button" name="restore" id="'.$data->id.'" class="btn btn-sm btn-success mr-1 mb-1 restore" data-toggle="tooltip" title="Restore data"><i class="fa fa-backward"></i></button>';
                     $trash .=' <button type="button" name="delete" id="'.$data->id.'" class="btn btn-sm btn-danger mr-1 mb-1 delete" data-toggle="tooltip" title="Destroy data"><i class="fa fa-trash-alt"></i></button>';
 
-                    if (request('trash') == 1){
+                    if (request('trash') == 1 && Gate::allows('role_delete')){
                         return $trash;
                     }else{
                         return $button;
@@ -59,7 +66,7 @@ class RoleController extends Controller
 
     public function create()
     {
-//        abort_if(Gate::denies('role_create'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(! Gate::allows('role_create'),403);
 
         $permissions = Permission::all()->pluck('name', 'id');
 
@@ -68,7 +75,7 @@ class RoleController extends Controller
 
     public function store(RolesStoreRequest $request)
     {
-
+        abort_if(! Gate::allows('role_create'),403);
         $role = Role::create(['name'=> $request->name,'description'=> $request->description]);
         $role->givePermissionTo($request->input('permissions'));
 
@@ -77,18 +84,16 @@ class RoleController extends Controller
 
     public function edit($id)
     {
-//        abort_if(Gate::denies('role_edit'), Response::HTTP_FORBIDDEN, '403 Forbidden');
-
+        abort_if(! Gate::allows('role_update'),403);
         $permissions = Permission::all()->pluck('name', 'id');
         $role=Role::findOrFail($id);
-//        $role->load('permissions');
 
         return view('admin.user_managements.roles.edit', compact('permissions', 'role'));
     }
 
     public function update(RolesUpdateRequest $request, $id)
     {
-
+        abort_if(! Gate::allows('role_update'),403);
         $role=Role::findOrFail($id);
         $role->update(['name'=> $request->name,'description'=> $request->description]);
         $role->syncPermissions($request->input('permissions'));
@@ -99,7 +104,7 @@ class RoleController extends Controller
 
     public function show($id)
     {
-//        abort_if(Gate::denies('role_show'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(! Gate::allows('role_show'),403);
 
         $role=Role::findOrFail($id);
         return view('admin.user_managements.roles.show', compact('role'));
@@ -107,27 +112,25 @@ class RoleController extends Controller
 
     public function destroy($id)
     {
-//        abort_if(Gate::denies('role_delete'), Response::HTTP_FORBIDDEN, '403 Forbidden');
+        abort_if(! Gate::allows('role_delete'),403);
         $role = Role::findOrFail($id);
         $role->delete();
 
         return response(__('role.delete_success'));
-
     }
 
     public function per_del($id)
     {
-
+        abort_if(! Gate::allows('role_delete'),403);
         $role=Role::onlyTrashed()->findOrFail($id);
         $role->forceDelete();
 
         return response(__('user.role_delete_success'));
-
     }
 
     public function restore($id)
     {
-
+        abort_if(! Gate::allows('role_delete'),403);
         $role=Role::onlyTrashed()->findOrFail($id);
         $role->restore();
